08 fe

+++ src/pages/PostDetail/PostDetail.js
@@ -12,6 +12,7 @@ import FavoriteIcon from "@material-ui/icons/Favorite";
 import MoodBadIcon from '@material-ui/icons/MoodBad';
 import Tooltip from '@material-ui/core/Tooltip';
 import Typography from '@material-ui/core/Typography';
+import FavoriteBorderIcon from '@material-ui/icons/FavoriteBorder';
 import ExpandMoreIcon from '@material-ui/icons/ExpandMore';
 import Accordion from '@material-ui/core/Accordion';
 import AccordionSummary from '@material-ui/core/AccordionSummary';
@@ -26,7 +27,6 @@ export const PostDetail = (props) => {
   const [post, setPost] = useState([]);
   const classes = useStyles();
   const [liked, setLiked] = useState(false)
-  // const [force, setForce] = useState(false)
   useEffect(() => {
     FetchData(
       `https://blog-fullstack-backend.herokuapp.com/post-detail/${slug}`
@@ -35,7 +35,6 @@ export const PostDetail = (props) => {
         setPost(results)
         isLiked(results)
       })
-      .catch((err) => console.log({ err }));
   }, [force]);
 
   const isLiked = (postDetail) =>{
@@ -46,10 +45,13 @@ export const PostDetail = (props) => {
                   setLiked(item.pk)
               }
           })
-      }).catch((err)=>console.log(err))
+      })
   }
 
   const handleLike = () =>{
+    if(!Authorization){
+      return null
+    }
     if(liked){
         axios.delete(`https://blog-fullstack-backend.herokuapp.com/like-detail/${post.slug}/${liked}/`,
         {
@@ -61,7 +63,6 @@ export const PostDetail = (props) => {
           setLiked(false)
           setForce(s=>!s)
         })
-        .catch((err)=>console.log(err))
         
     }else{
         axios.post(`https://blog-fullstack-backend.herokuapp.com/like/${post.slug}/`,{
@@ -73,16 +74,11 @@ export const PostDetail = (props) => {
             }
         })
         .then((msg)=>setForce(s=>!s))
-        .catch((err)=>console.log(err))
         setLiked(true)
         setForce(s=>!s)
     }
 }
 
-  console.log({post});
-  console.log({liked})
-  // console.log(post.user == currentUser);
-  // console.log(post.slug, post.pk);
   return (
     <Grid container xs={12} justify="center" className={classes.root}>
       <Grid item>
@@ -103,10 +99,10 @@ export const PostDetail = (props) => {
             <div className={classes.iconButtons}>
               <IconButton onClick={handleLike}>
                 <Badge badgeContent={post?.like_count} color="secondary">
-                  <FavoriteIcon color={liked ? "secondary" : "action"} />
+                  {liked ? <FavoriteIcon color="secondary" /> : <FavoriteBorderIcon color="primary" />}
                 </Badge>
               </IconButton>
-              <IconButton aria-label="add to favorites">
+              <IconButton>
                 <Tooltip title="Report as Bad Post" arrow>
                   <MoodBadIcon color="primary" />
                 </Tooltip>
